Working model
v0.1.0
A basic command line version of the game.

The secrect word is fixed = JAVA.

The player can continue guessing up to max of 3 incorrect guesses (LOSES) or guesses all the characters in the secrect word (WINS).

Different messages are printed to console, e.g., correct/incorrect guess, number of incorrect guesses, game over etc.

The WordProcessor class has all the functions to check if a guessed character is correct, fill in the blanks of the current guessed word, and check if all characters have been guessed.

inputValidation is still yet to be implemented.